Migrate from modules to crates
Previously all runtimes were united in one crate and toggled by using flags. The client must be able to offer all runtimes, which was impossible with all runtimes being in one crate: Rust prohibits linking to the same crate with different features. Consequently, each runtime must have its own crate. This complicates things significantly and if the previously common runtime schema wants to be used, wrapper macros have to be introduced.
