--- runtime/src/common/benchmarks/mod.rs
@@ -1,6 +1,0 @@
-#![cfg(feature = "runtime-benchmarks")]
-#![allow(clippy::integer_arithmetic)]
-
-pub mod currencies;
-pub mod tokens;
-pub mod utils;
