--- runtime/src/common/tests/mod.rs
@@ -1,3 +1,0 @@
-#![cfg(test)]
-
-mod multiplier;
