+++ runtime/zeitgeist/build.rs
@@ -1,0 +1,5 @@
+use substrate_wasm_builder::WasmBuilder;
+
+fn main() {
+    WasmBuilder::new().with_current_project().export_heap_base().import_memory().build()
+}
