+++ runtime/zeitgeist/src/lib.rs
@@ -1,16 +1,67 @@
-#![cfg(not(feature = "testnet"))]
+#![cfg_attr(not(feature = "std"), no_std)]
+#![recursion_limit = "256"]
 
-use super::common::*;
+extern crate alloc;
+
+#[cfg(feature = "std")]
+include!(concat!(env!("OUT_DIR"), "/wasm_binary.rs"));
+
+use common_runtime::{create_runtime_with_additional_pallets, impl_config_traits, create_runtime_apis, decl_common_types, create_common_benchmark_logic, create_common_tests};
 pub use frame_system::{
     Call as SystemCall, CheckEra, CheckGenesis, CheckNonZeroSender, CheckNonce, CheckSpecVersion,
     CheckTxVersion, CheckWeight,
 };
 pub use pallet_transaction_payment::ChargeTransactionPayment;
-pub use parameters::*;
 #[cfg(feature = "parachain")]
-pub use {pallet_author_slot_filter::EligibilityValue, parachain_params::*};
+pub use {pallet_author_slot_filter::EligibilityValue};
 
-use frame_support::{construct_runtime, traits::Contains};
+// Expose Runtime
+pub use {api, parameters::SS58Prefix, Call, Runtime, RuntimeApi};
+#[cfg(feature = "std")]
+pub use zeitgeist::{
+    AdvisoryCommitteeMembershipConfig, BalancesConfig, CouncilMembershipConfig, GenesisConfig,
+    LiquidityMiningConfig, SystemConfig, TechnicalCommitteeMembershipConfig,
+};
+#[cfg(all(feature = "std", not(feature = "parachain")))]
+pub use zeitgeist::{AuraConfig, GrandpaConfig};
+
+// Expose functions and types required to construct node CLI
+#[cfg(feature = "std")]
+pub use common::native_version;
+pub use common::{
+    opaque::Block, ChargeTransactionPayment, CheckEra, CheckGenesis, CheckNonZeroSender,
+    CheckNonce, CheckSpecVersion, CheckTxVersion, CheckWeight, SignedExtra, SignedPayload,
+    SystemCall, UncheckedExtrinsic,
+};
+
+
+use alloc::vec;
+use frame_support::{
+    traits::{ConstU16, ConstU32, Contains, EnsureOneOf, EqualPrivilegeOnly, InstanceFilter},
+    weights::{constants::RocksDbWeight, ConstantMultiplier, IdentityFee},
+};
+use frame_system::EnsureRoot;
+use pallet_collective::{EnsureProportionAtLeast, PrimeDefaultVote};
+use sp_runtime::{
+    generic,
+    traits::{AccountIdConversion, AccountIdLookup, BlakeTwo256},
+};
+#[cfg(feature = "std")]
+use sp_version::NativeVersion;
+use substrate_fixed::{types::extra::U33, FixedI128, FixedU128};
+use zeitgeist_primitives::{constants::*, types::*};
+use zrml_rikiddo::types::{EmaMarketVolume, FeeSigmoid, RikiddoSigmoidMV};
+#[cfg(feature = "parachain")]
+use {
+    frame_support::traits::{Everything, Nothing},
+    frame_system::EnsureSigned,
+    xcm_builder::{EnsureXcmOrigin, FixedWeightBounds, LocationInverter},
+    xcm_config::XcmConfig,
+};
+
+
+
+use frame_support::{construct_runtime};
 
 use sp_api::impl_runtime_apis;
 use sp_core::{crypto::KeyTypeId, OpaqueMetadata};
@@ -26,8 +77,10 @@ use nimbus_primitives::{CanAuthor, NimbusId};
 use sp_version::RuntimeVersion;
 use zeitgeist_primitives::types::*;
 
-pub mod parachain_params;
-pub mod parameters;
+#[cfg(feature = "parachain")]
+mod xcm_config;
+mod parachain_params;
+mod parameters;
 
 pub const VERSION: RuntimeVersion = RuntimeVersion {
     spec_name: create_runtime_str!("zeitgeist"),
@@ -74,5 +127,9 @@ impl Contains<Call> for IsCallable {
     }
 }
 
+decl_common_types!();
 create_runtime_with_additional_pallets!();
+impl_config_traits!();
 create_runtime_apis!();
+create_common_benchmark_logic!();
+create_common_tests!();
